Add newCells and replaceCells as options
Currently new code cells are not detected and each time the page changes the bootstrap function needs to be executed to detect new code cells. This add 2 new function to either detect new code cells and add them to the list or replace the old list with the new code cells.